Government_Scheme_Summary_Query_using_LangChain_OpenAI/requirements.txt
ipykernel
python-dotenv
langchain
openai
langchain-openai
langchain_community
requests
pdfplumber
faiss-cpu
bs4
streamlit